rename : rename test class and move directory
