+++ src/test/java/nextstep/courses/domain/session/file/150w.png
