+++ src/test/java/nextstep/courses/domain/session/file/1mb.jpg
