+++ src/test/java/nextstep/courses/domain/session/file/image.tif
