+++ src/test/java/nextstep/courses/domain/session/file/invalid_ratio.png
